feat(web): add tabs to export page
