+++ packages_rs/nextclade-web/src/components/FilePicker/FilePicker.tsx
@@ -6,7 +6,7 @@ import styled from 'styled-components'
 import { useTranslationSafe } from 'src/helpers/useTranslationSafe'
 import type { AlgorithmInput } from 'src/types'
 import { AlgorithmInputFile, AlgorithmInputString, AlgorithmInputUrl } from 'src/io/AlgorithmInput'
-import { TabsContent, TabsPanel } from 'src/components/Common/Tabs'
+import { TabsContent, TabsPanel } from 'src/components/Common/TabsLight'
 import { UploadedFileLoadingInfo } from 'src/components/FilePicker/UploadedFileLoadingInfo'
 import { UploadBox } from './UploadBox'
 import { UploadBoxCompact } from './UploadBoxCompact'
